requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorlog==6.8.2
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
deepl==1.16.1
defusedxml==0.7.1
diffusers==0.26.0
dill==0.3.7
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gdown==5.1.0
gensim==4.3.2
grad-cam==1.5.0
greenlet==3.0.3
huggingface-hub==0.20.2
idna==3.6
imageio==2.34.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy_loader==0.4
Mako==1.3.2
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.1
nlpaug==1.1.11
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
opencv-python==4.9.0.80
optuna==3.6.1
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.20.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge_score==0.1.2
rpds-py==0.17.1
sacremoses==0.1.1
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.15.0
tomli==2.0.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
triton==2.1.0
ttach==0.0.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
